Simplify stats collection
All arguments passed to recordCallSite are now joined on space and
stored, no attempt is made to combine them.

Rolling up the information (summing for counts, min/max/mean/... for
other numbers) will be done after the fact and that's up to the user.

This class does not make any assumptions about the information that is
being recorded for eac call site. It could be hit counts, run times,
strings, ...