Make everything static
It makes more sense for these functions which are supposed to record
values for aggregation be static instead of instance. This is
particularly true for stats collection. It would be annoying to have to
collect and aggregate stats across many instances of a class, collecting
the results statically makes more sense.

Though, PHP being the odd language it is, still allows you to call
static methods in an instancey way: `$this->someStaticMethod()` is
valid.